Remove SQL dump handling from Docker setup and update WordPress installation script with environment variables

+++ srcs/requirements/wordpress/docker-entrypoint.sh
@@ -6,9 +6,25 @@ while ! mysql -u ${MYSQL_USER} -h ${MYSQL_HOST} -p${MYSQL_PASSWORD} ${MYSQL_DATA
     sleep 1
 done
 
-echo Installing redis-cache plugin...
-wp plugin install redis-cache --activate
-wp redis enable
+if [ -e /var/www/wordpress/wp-config-sample.php ]; then
+	echo installing wordpress...
+    wp core install --url=$WORDPRESS_DOMAIN/ \
+                        --title=$WORDPRESS_TITLE\
+                        --admin_user=$WORDPRESS_ADMIN_USER\
+                        --admin_password=$WORDPRESS_ADMIN_PASSWORD\
+                        --admin_email=$WORDPRESS_ADMIN_EMAIL\
+                        --skip-email\
+                        --allow-root
+    wp user create    $WORDPRESS_USER\
+                        $WORDPRESS_EMAIL\
+                        --user_pass=$WORDPRESS_PASSWORD\
+                        --role=author\
+                        --allow-root
+	echo Installing redis-cache plugin...
+	wp plugin install redis-cache --activate
+	wp redis enable
+	rm /var/www/wordpress/wp-config-sample.php
+fi
 
 echo Starting wordpress...
 php-fpm83 -F
